requirements.txt
Cython==0.25.2
PyVCF==0.6.8
